First pass at fixing 2563.
