+++ core/src/main/kotlin/net/corda/core/node/services/KeyManagementService.kt
@@ -11,6 +11,7 @@ import java.security.KeyPair
 import java.security.PrivateKey
 import java.security.PublicKey
 import java.security.cert.X509Certificate
+import java.util.*
 
 /**
  * The KMS is responsible for storing and using private keys to sign things. An implementation of this may, for example,
@@ -31,6 +32,13 @@ interface KeyManagementService {
     fun freshKey(): PublicKey
 
     /**
+     * Generates a new random [KeyPair] and adds it to the internal key storage. Associates the public key to an external ID. Returns the
+     * public part of the pair.
+     */
+    @Suspendable
+    fun freshKey(externalId: UUID): PublicKey
+
+    /**
      * Generates a new random [KeyPair], adds it to the internal key storage, then generates a corresponding
      * [X509Certificate] and adds it to the identity service.
      *
@@ -40,6 +48,18 @@ interface KeyManagementService {
      */
     @Suspendable
     fun freshKeyAndCert(identity: PartyAndCertificate, revocationEnabled: Boolean): PartyAndCertificate
+
+    /**
+     * Generates a new random [KeyPair], adds it to the internal key storage, then generates a corresponding
+     * [X509Certificate] and adds it to the identity service.
+     *
+     * @param identity identity to generate a key and certificate for. Must be an identity this node has CA privileges for.
+     * @param revocationEnabled whether to check revocation status of certificates in the certificate path.
+     * @param externalId ID to associate the newly created [PublicKey] with.
+     * @return X.509 certificate and path to the trust root.
+     */
+    @Suspendable
+    fun freshKeyAndCert(identity: PartyAndCertificate, revocationEnabled: Boolean, externalId: UUID): PartyAndCertificate
 
     /**
      * Filter some keys down to the set that this node owns (has private keys for).
